server/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
dlib==19.24.2
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.6
numpy==1.26.3
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
zope.interface==6.2
